requirements.txt
dbus-python==1.2.8
mpris2==1.0.2
